* Reorganize components and css folder structure.
