+++ app/assets/src/components/PipelineSampleReads.jsx
@@ -3,9 +3,8 @@ import ReactDOM from "react-dom";
 import moment from "moment";
 import $ from "jquery";
 import axios from "axios";
-import Cookies from "js-cookie";
-import { Button, Icon, Divider, Popup, Dropdown } from "semantic-ui-react";
-import DownloadButton from "./controls/DownloadButton";
+import { Button, Divider, Dropdown } from "semantic-ui-react";
+import DownloadButton from "./ui/controls/buttons/DownloadButton";
 import numberWithCommas from "../helpers/strings";
 import SubHeader from "./SubHeader";
 import ERCCScatterPlot from "./ERCCScatterPlot";
