+++ app/assets/src/components/ui/controls/Slider.jsx
