+++ app/assets/src/components/ui/icons/CompareIcon.jsx
