+++ app/assets/src/components/ui/icons/RemoveIcon.jsx
