+++ app/assets/src/styles/ui/controls/dropdowns/multiple_dropdown.scss
