+++ app/assets/src/styles/ui/controls/dropdowns/threshold_filter_dropdown.scss
