+++ app/assets/src/styles/ui/controls/slider.scss
